requirements.txt
asgiref==3.4.0
chess==1.6.1
dj-database-url==0.5.0
Django==3.2.4
django-filter==2.4.0
django-heroku==0.3.1
django-user-agents==0.4.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
psycopg2==2.9.1
python-chess==1.999
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
ua-parser==0.10.0
user-agents==2.2.0
whitenoise==5.2.0
